tools/developer_tools/bely-mqtt-message-broker/pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bely-mqtt-framework"
version = "0.1.0"
description = "Pluggable Python framework for handling BELY MQTT events"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "BELY Team", email = "team@bely.dev"}
]
keywords = ["mqtt", "bely", "logbook", "plugins", "framework"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "click>=8.1.0",
    "paho-mqtt>=1.6.1",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "pluggy>=1.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "types-paho-mqtt>=1.6.0",
]
apprise = [
    "apprise>=1.4.0",
]

[project.scripts]
bely-mqtt = "bely_mqtt.cli:main"

[tool.black]
line-length = 100
target-version = ['py39']

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
